Rename AbstractInterfaceManager to WebviewBase
